style adjustments
convert to a more c-like style where reference usage is limited (but not
completely eliminated). this adds the benefit of having everything being
default constructable and so we get very nice composability. there are a
few drawbacks like having to initialize sentinel-based linked lists
in-place, but that's not a huge concern.